moved write_config_file to utilities.py
